--- a/blake3.go
+++ b/blake3.go
@@ -45,41 +45,31 @@ func wordsToBytes(words []uint32, bytes []byte) {
 	}
 }
 
-// The g function, split into two parts so that the compiler will inline it.
-func gx(state *[16]uint32, a, b, c, d int, mx uint32) {
-	state[a] += state[b] + mx
-	state[d] = bits.RotateLeft32(state[d]^state[a], -16)
-	state[c] += state[d]
-	state[b] = bits.RotateLeft32(state[b]^state[c], -12)
-}
-
-func gy(state *[16]uint32, a, b, c, d int, my uint32) {
-	state[a] += state[b] + my
-	state[d] = bits.RotateLeft32(state[d]^state[a], -8)
-	state[c] += state[d]
-	state[b] = bits.RotateLeft32(state[b]^state[c], -7)
+// The g function
+func g(a, b, c, d, mx, my uint32) (uint32, uint32, uint32, uint32) {
+	a += b + mx
+	d = bits.RotateLeft32(d^a, -16)
+	c += d
+	b = bits.RotateLeft32(b^c, -12)
+	a += b + my
+	d = bits.RotateLeft32(d^a, -8)
+	c += d
+	b = bits.RotateLeft32(b^c, -7)
+	return a, b, c, d
 }
 
 func round(state *[16]uint32, m *[16]uint32) {
 	// Mix the columns.
-	gx(state, 0, 4, 8, 12, m[0])
-	gy(state, 0, 4, 8, 12, m[1])
-	gx(state, 1, 5, 9, 13, m[2])
-	gy(state, 1, 5, 9, 13, m[3])
-	gx(state, 2, 6, 10, 14, m[4])
-	gy(state, 2, 6, 10, 14, m[5])
-	gx(state, 3, 7, 11, 15, m[6])
-	gy(state, 3, 7, 11, 15, m[7])
+	state[0], state[4], state[8], state[12] = g(state[0], state[4], state[8], state[12], m[0], m[1])
+	state[1], state[5], state[9], state[13] = g(state[1], state[5], state[9], state[13], m[2], m[3])
+	state[2], state[6], state[10], state[14] = g(state[2], state[6], state[10], state[14], m[4], m[5])
+	state[3], state[7], state[11], state[15] = g(state[3], state[7], state[11], state[15], m[6], m[7])
 
 	// Mix the diagonals.
-	gx(state, 0, 5, 10, 15, m[8])
-	gy(state, 0, 5, 10, 15, m[9])
-	gx(state, 1, 6, 11, 12, m[10])
-	gy(state, 1, 6, 11, 12, m[11])
-	gx(state, 2, 7, 8, 13, m[12])
-	gy(state, 2, 7, 8, 13, m[13])
-	gx(state, 3, 4, 9, 14, m[14])
-	gy(state, 3, 4, 9, 14, m[15])
+	state[0], state[5], state[10], state[15] = g(state[0], state[5], state[10], state[15], m[8], m[9])
+	state[1], state[6], state[11], state[12] = g(state[1], state[6], state[11], state[12], m[10], m[11])
+	state[2], state[7], state[8], state[13] = g(state[2], state[7], state[8], state[13], m[12], m[13])
+	state[3], state[4], state[9], state[14] = g(state[3], state[4], state[9], state[14], m[14], m[15])
 }
 
 func permute(m, n *[16]uint32) {