* avoid copying on permute.
* add a capacity hint on bytesToWords.
* merge gx and gy into g.
```
name      old time/op    new time/op    delta
Write-4     4.62µs ± 1%    4.04µs ± 0%  -12.61%  (p=0.000 n=43+9)
Sum256-4    4.59µs ± 1%    3.96µs ± 0%  -13.75%  (p=0.000 n=41+9)
XOF-4       4.34µs ± 0%    3.79µs ± 0%  -12.55%  (p=0.000 n=44+9)

name      old speed      new speed      delta
Write-4    221MB/s ± 1%   253MB/s ± 0%  +14.43%  (p=0.000 n=43+9)
Sum256-4   223MB/s ± 1%   259MB/s ± 0%  +15.94%  (p=0.000 n=41+9)
XOF-4      236MB/s ± 0%   270MB/s ± 0%  +14.38%  (p=0.000 n=44+8)

name      old alloc/op   new alloc/op   delta
Write-4      0.00B          0.00B          ~     (all equal)
Sum256-4     0.00B          0.00B          ~     (all equal)
XOF-4        0.00B          0.00B          ~     (all equal)

name      old allocs/op  new allocs/op  delta
Write-4       0.00           0.00          ~     (all equal)
Sum256-4      0.00           0.00          ~     (all equal)
XOF-4         0.00           0.00          ~     (all equal)
```